+++ AMaze/ScanIntersectionExtra.cs
@@ -1,7 +1,0 @@
-﻿namespace AMaze;
-
-internal struct ScanIntersectionExtra
-{
-	public double top, bottom;
-	public bool opaque, altPalette, vantablack;
-}